1.修改职位推荐UI切换 底部留白 2.修改职位推荐显示金额 3.修改职位详情页面网络请求接口 3.修改职位详情UI页面

+++ app/src/main/java/com/tntlinking/tntdev/http/api/GetDeveloperRecommendsApi.java
@@ -16,8 +16,8 @@ public final class GetDeveloperRecommendsApi implements IRequestApi {
         private String positionId;
         private String title;
         private String workDaysModeName;
-        private int startPay;
-        private int endPay;
+        private double startPay;
+        private double endPay;
         private String educationName;
         private String workYearsName;
 
@@ -30,6 +30,7 @@ public final class GetDeveloperRecommendsApi implements IRequestApi {
         }
 
         private Boolean recommendByOperate;
+
         public List<String> getSkillNames() {
             return skillNames;
         }
@@ -69,19 +70,19 @@ public final class GetDeveloperRecommendsApi implements IRequestApi {
             this.workDaysModeName = workDaysModeName;
         }
 
-        public int getStartPay() {
+        public double getStartPay() {
             return startPay;
         }
 
-        public void setStartPay(int startPay) {
+        public void setStartPay(double startPay) {
             this.startPay = startPay;
         }
 
-        public int getEndPay() {
+        public double getEndPay() {
             return endPay;
         }
 
-        public void setEndPay(int endPay) {
+        public void setEndPay(double endPay) {
             this.endPay = endPay;
         }
